nuked the database, replaced the addressId with int, fixed the create gardener method so its fully functioning, updated the create view for gardener.

+++ CommunityGardenProj/Models/Address.cs
@@ -9,7 +9,8 @@ namespace CommunityGardenProj.Models
     public class Address
     {
         [Key]
-        public double AddressId { get; set; }
+        //public int Id { get; set; }
+        public int AddressId { get; set; }
 
         public string StreetAddress { get; set; }
         public string City { get; set; }
